Rename community folders to home
