+++ src/modules/home/components/AppPageLayout.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 
 import PageLayout from '@/modules/shared/components/PageLayout';
-import PageStructure from '@/modules/community/components/AppPageStructure';
+import PageStructure from '@/modules/home/components/AppPageStructure';
 
 export function getCommunityLayout(page: ReactElement) {
   return (
