+++ src/modules/home/helpers/appInfo.ts
