+++ src/modules/home/styles/app.imports.js
@@ -3,7 +3,7 @@
 /* Does not include app.main.js styles and *.module.css styles */
 
 /* Variables */
-import colors from '@/modules/community/styles/var/colors.var';
+import colors from '@/modules/home/styles/var/colors.var';
 
 export {
   colors,
